requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
nltk==3.2.4
oauthlib==3.1.1
preprocessing==0.1.13
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sphinx-rtd-theme==0.2.4
uritemplate==3.0.1
urllib3==1.26.7
